optimization of the code
